Refactor :  좋아요 컴포넌트 더미데이터 추가 및 적용 후 재사용 가능하도록 UI 컴포넌트로 분리
